voice-guard/api/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Google Cloud Services
google-cloud-speech==2.21.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-genai


# WebSocket & File Upload
websockets==12.0
python-multipart==0.0.6

# Environment & Data Validation (Optional)
python-dotenv==1.0.0
pydantic==2.5.0
